refactor(test page): collapse section titles into ListView of tiles
TestPage 之前每个 section 都是 '粗体小标题 + Card(ListTile)' 的两层
结构，标题文字在 tile 上面、tile 自己的 title 又写一遍，重复且
浪费垂直空间。改成 ListView 堆叠裸的 ListTile 后变成 Android 系统
设置/About 页的标准形态。

具体改动：

1. 三个 *button widget 改名为 *tile widget，去掉外层 Card，直接
   返回 ListTile：
   - environment_info_button.dart  -> environment_info_tile.dart
     (EnvironmentInfoButton -> EnvironmentInfoTile)
   - wizard_reset_button.dart     -> wizard_reset_tile.dart
     (WizardResetButton -> WizardResetTile)
   - auth_log/auth_log_card.dart  -> auth_log/auth_log_tile.dart
     (AuthLogCard -> AuthLogTile)

2. 解决命名冲突：AuthLogTile 之前已经用于 '单条日志行 widget'，
   改名 AuthLogEntryTile 并单独放到 auth_log_entry_tile.dart，
   viewer 里的 ListView.builder 改用它。

3. test_page.dart 重构：
   - SingleChildScrollView + Column  -> ListView
   - 删除 _SectionTitle widget（三个 section 都去掉了小标题）
   - 删掉 32px SizedBox 分隔，ListTile 自然堆叠
   - Update section 保留自己的小标题（4 个子部件的独立 block）
   - 整个 body 加 EdgeInsets.all(24) 外层 padding（ListTile 自带
     水平 contentPadding，所以最终水平间距是 24+16=40px，跟原
     SingleChildScrollView 的视觉效果一致）

4. 删除遗留的 lib/pages/test/auth_log_viewer_page.dart（在上一
   次拆分子目录时漏删的旧文件，全工程已无引用）。
